move model to schemas module
